requirements.txt
albumentations==1.0.0
fastapi==0.66.0
# glob2==0.7
# imgaug==0.4.0
# jupyterlab==3.0.16
# levenshtein==0.12.0
# matplotlib==3.4.2
# notebook==6.4.0
numpy==1.20.3
opencv-python==4.5.2.52
# pandas==1.2.4
Pillow==8.2.0
# scikit-image==0.18.1
# scikit-learn==0.24.2
# scipy==1.6.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cpu
# torchsummary==1.5.1
# tqdm==4.60.0
uvicorn==0.14.0
# wandb==0.10.30